refactor: rename log drivers API
