+++ libs/ngworker/lumberjack/console-driver/src/log-drivers-api.spec.ts
@@ -2,7 +2,7 @@ import { isClass } from '@internal/test-util';
 
 import { ConsoleDriver } from './index';
 
-describe('Log driver API', () => {
+describe('Log drivers API', () => {
   describe('Services', () => {
     it(`exposes ${ConsoleDriver.name}`, () => {
       const sut = ConsoleDriver;
